fix: improve preflight check performance and reliability
Replace full LLM dry-run with lightweight metric instantiation check,
wrap blocking sync calls in asyncio.to_thread(), add global timeout
with failure event publishing, and refactor into a module package.
