+++ apps/backend/src/rhesis/backend/app/services/preflight/__init__.py
@@ -1,0 +1,29 @@
+"""Preflight check service for validating test execution environment."""
+
+from .constants import (
+    CHECK_BEHAVIOR_METRIC_COVERAGE,
+    CHECK_ENDPOINT_CONNECTIVITY,
+    CHECK_EVALUATION_MODEL,
+    CHECK_EXECUTION_MODEL,
+    CHECK_METRIC_FUNCTIONALITY,
+    CHECK_TEST_SET_NOT_EMPTY,
+    LABELS,
+    PER_TEST_SET_CHECKS,
+    SHARED_CHECKS,
+)
+from .orchestrator import compute_summary, run_preflight_checks, run_preflight_checks_multi
+
+__all__ = [
+    "CHECK_BEHAVIOR_METRIC_COVERAGE",
+    "CHECK_ENDPOINT_CONNECTIVITY",
+    "CHECK_EVALUATION_MODEL",
+    "CHECK_EXECUTION_MODEL",
+    "CHECK_METRIC_FUNCTIONALITY",
+    "CHECK_TEST_SET_NOT_EMPTY",
+    "LABELS",
+    "PER_TEST_SET_CHECKS",
+    "SHARED_CHECKS",
+    "compute_summary",
+    "run_preflight_checks",
+    "run_preflight_checks_multi",
+]
